+++ roncoo-education-app-gateway/src/main/java/com/roncoo/education/ServerGatewayApplication.java
@@ -3,6 +3,7 @@
  */
 package com.roncoo.education;
 
+import com.roncoo.education.app.gateway.common.FilterPre;
 import org.springframework.boot.SpringApplication;
 import org.springframework.cloud.client.SpringCloudApplication;
 import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
@@ -10,8 +11,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.filter.CorsFilter;
-
-import com.roncoo.education.server.gateway.common.FilterPre;
 
 /**
  * 服务网关
@@ -22,26 +21,26 @@ import com.roncoo.education.server.gateway.common.FilterPre;
 @SpringCloudApplication
 public class ServerGatewayApplication {
 
-	@Bean
-	public CorsFilter corsFilter() {
-		final UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-		final CorsConfiguration config = new CorsConfiguration();
-		config.setAllowCredentials(true);
-		config.addAllowedOrigin("*");
-		config.addAllowedHeader("*");
-		config.addAllowedMethod("*");
-		config.setMaxAge(18000L);
-		source.registerCorsConfiguration("/**", config);
-		return new CorsFilter(source);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ServerGatewayApplication.class, args);
+    }
 
-	public static void main(String[] args) {
-		SpringApplication.run(ServerGatewayApplication.class, args);
-	}
+    @Bean
+    public FilterPre filterPre() {
+        return new FilterPre();
+    }
 
-	@Bean
-	public FilterPre filterPre() {
-		return new FilterPre();
-	}
+    @Bean
+    public CorsFilter corsFilter() {
+        final UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        final CorsConfiguration config = new CorsConfiguration();
+        config.setAllowCredentials(true);
+        config.addAllowedOrigin("*");
+        config.addAllowedHeader("*");
+        config.addAllowedMethod("*");
+        config.setMaxAge(18000L);
+        source.registerCorsConfiguration("/**", config);
+        return new CorsFilter(source);
+    }
 
 }
