+++ roncoo-education-app-gateway/src/main/resources/bootstrap.properties
@@ -1,13 +1,13 @@
 # application
 spring.application.name=roncoo-education-app-gateway
 # profile
-spring.profiles.active=dev
+spring.profiles.active=demo
 # server
 server.port=5840
 # nacos
 spring.cloud.nacos.username=nacos
-spring.cloud.nacos.password=nacos
-spring.cloud.nacos.server-addr=localhost:8848
+spring.cloud.nacos.password=RonCoo.123
+spring.cloud.nacos.server-addr=10.65.3.98:8848
 spring.cloud.nacos.namespace=${spring.profiles.active}
 spring.cloud.nacos.discovery.namespace=${spring.cloud.nacos.namespace}
 spring.cloud.nacos.config.namespace=${spring.cloud.nacos.namespace}
