+++ roncoo-education-app-gateway/src/main/resources/templates/index.ftl
@@ -1,19 +1,19 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-  <meta charset="UTF-8">
-  <title>接口文档</title>
+    <meta charset="UTF-8">
+    <title>接口文档</title>
 </head>
 <body>
-  <div style="margin:0 auto;text-align:center;">
+<div style="margin:0 auto;text-align:center;">
     <br/>
-    <a href="http://192.168.1.75:5710/doc.html" target="_blanck">课程API</a>
-    <br/>
-    <br/>
-    <a href="http://192.168.1.75:5720/doc.html" target="_blanck">用户API</a>
+    <a href="http://localhost:5710/doc.html" target="_blanck">课程API</a>
     <br/>
     <br/>
-    <a href="http://192.168.1.75:5730/doc.html" target="_blanck">系统API</a>
-  </div>
+    <a href="http://localhost:5720/doc.html" target="_blanck">用户API</a>
+    <br/>
+    <br/>
+    <a href="http://localhost:5730/doc.html" target="_blanck">系统API</a>
+</div>
 </body>
 </html>
