+++ roncoo-education-app-sba/src/main/java/com/roncoo/education/app/sba/ServerSbaApplication.java
@@ -1,21 +1,20 @@
 /**
  * Copyright 2015-现在 广州市领课网络科技有限公司
  */
-package com.roncoo.education.server.sba;
+package com.roncoo.education.app.sba;
 
+import de.codecentric.boot.admin.server.config.EnableAdminServer;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-import de.codecentric.boot.admin.server.config.EnableAdminServer;
 
 @EnableAdminServer
 @EnableDiscoveryClient
 @SpringBootApplication
 public class ServerSbaApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ServerSbaApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ServerSbaApplication.class, args);
+    }
 
 }
